WIP parallel progress formatter
This is an attempt to create a parallel progress formatter for rspec.
The idea is that the each rspec process's formatter connects to a DRb
server in the parallel_tests parent process, posting events to it. I
think. Unfortunately it currently relies on mods to parallel_tests
itself, which is extremely non-ideal.